flask_reporter/requirements.txt
Flask
APScheduler
psycopg2-binary